DRY up the rfqt mocker
