+++ packages/asset-swapper/src/index.ts
@@ -86,7 +86,7 @@ export {
     MarketBuySwapQuote,
     MarketOperation,
     MarketSellSwapQuote,
-    MockedRfqtFirmQuoteResponse,
+    MockedRfqtQuoteResponse,
     OrderPrunerPermittedFeeTypes,
     OrderWithFillableAmounts,
     RfqtMakerAssetOfferings,
@@ -164,7 +164,7 @@ export {
     QuoteReportEntry,
 } from './utils/quote_report_generator';
 export { QuoteRequestor } from './utils/quote_requestor';
-export { rfqtMocker } from './utils/rfqt_mocker';
+export { testHelpers, RfqtQuoteEndpoint } from '../test/utils/test_helpers';
 export { ERC20BridgeSamplerContract, BalanceCheckerContract } from './wrappers';
 import { ERC20BridgeSource } from './utils/market_operation_utils/types';
 export type Native = ERC20BridgeSource.Native;
