+++ packages/asset-swapper/test/utils/test_helpers.ts
@@ -1,6 +1,14 @@
 import { BigNumber } from '@0x/utils';
+import axios, { AxiosInstance } from 'axios';
+import AxiosMockAdapter from 'axios-mock-adapter';
 
 import { InsufficientAssetLiquidityError } from '../../src/errors';
+import { MockedRfqtQuoteResponse } from '../../src/types';
+
+export enum RfqtQuoteEndpoint {
+    Indicative = 'price',
+    Firm = 'quote',
+}
 
 export const testHelpers = {
     expectInsufficientLiquidityErrorAsync: async (
@@ -22,5 +30,32 @@ export const testHelpers = {
         }
 
         expect(wasErrorThrown).to.be.true();
+    },
+    /**
+     * A helper utility for testing which mocks out
+     * requests to RFQ-t providers
+     */
+    withMockedRfqtQuotes: async (
+        mockedResponses: MockedRfqtQuoteResponse[],
+        quoteType: RfqtQuoteEndpoint,
+        afterResponseCallback: () => Promise<void>,
+        axiosClient: AxiosInstance = axios,
+    ): Promise<void> => {
+        const mockedAxios = new AxiosMockAdapter(axiosClient);
+        try {
+            // Mock out RFQT responses
+            for (const mockedResponse of mockedResponses) {
+                const { endpoint, requestApiKey, requestParams, responseData, responseCode } = mockedResponse;
+                const requestHeaders = { Accept: 'application/json, text/plain, */*', '0x-api-key': requestApiKey };
+                mockedAxios
+                    .onGet(`${endpoint}/${quoteType}`, { params: requestParams }, requestHeaders)
+                    .replyOnce(responseCode, responseData);
+            }
+            // Perform the callback function, e.g. a test validation
+            await afterResponseCallback();
+        } finally {
+            // Ensure we always restore axios afterwards
+            mockedAxios.restore();
+        }
     },
 };
